+++ CHANGELOG.md
@@ -6,3 +6,10 @@
 
 * initial value for each doropdown added
 * Alignment fixed for selected item and a dropdown menu items
+
+## 0.0.5
+
+* autoSelectInitialValue option added 
+* bug fixing for label removal when adding custom decoration
+* dropdown made nullable
+* nepali locale for District fixed